Fastapi-queryservice/requirements.txt
# FastAPI ERP Query Service - Production Dependencies
# Core FastAPI framework
fastapi==0.116.1
uvicorn[standard]==0.33.0

# Database connectivity
sqlalchemy==2.0.43
mysql-connector-python==9.0.0

# Data validation and settings
pydantic==2.10.6
pydantic-core==2.27.2
pydantic-settings==2.8.1

# Environment and configuration
python-dotenv==1.0.1

# Performance enhancements
uvloop==0.21.0  # High-performance event loop (Linux/macOS)
httptools==0.6.4  # High-performance HTTP parser

# HTTP and async support
anyio==4.5.2
h11==0.16.0
starlette==0.44.0
click==8.1.8

# Type annotations and validation
annotated-types==0.7.0
typing-extensions==4.13.2

# Utilities
idna==3.10
sniffio==1.3.1
greenlet==3.1.1
exceptiongroup==1.3.0

# Development and monitoring (optional)
watchfiles==0.24.0
websockets==13.1
PyYAML==6.0.2

# Mobile API Testing Dependencies
pytest==7.4.4
pytest-asyncio==0.21.1
httpx==0.25.2
python-multipart==0.0.6
